docs: git SHA prefixes are for humans
Clarify that the preferred git SHA abbreviation length for the Fixes tag
is 12 characters[1], as the tag is intended for humans (though, yes,
it is parsed by machines too). When resolving collisions, the quoted
parenthetical patch Subject that follows the abbreviated hash should be
used to identify the intended commit.

Suggested-by: Linus Torvalds <torvalds@linux-foundation.org>
Link: https://lore.kernel.org/all/CAHk-=wiwAz3UgPOWK3RdGXDnTRHcwVbxpuxCQt_0SoAJC-oGXQ@mail.gmail.com [1]
Signed-off-by: Kees Cook <kees@kernel.org>